+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Calling `Dial()` with a cancelled context doesn't create a connection.
 * Context cancellation is properly honored in calls to `Dial()` and `NewConn()`.
 * Fixed potential race during `Conn.Close()`.
+* Disable sending frames when closing `Session`, `Sender`, and `Receiver`.
 
 ### Other Changes